Update wss_curves_areas_post.py
Landuse conversion default

+++ hhnk_research_tools/waterschadeschatter/wss_curves_areas_post.py
@@ -17,7 +17,9 @@ import pandas as pd
 import shapely
 from tqdm import tqdm
 
+import hhnk_research_tools as hrt
 import hhnk_research_tools.logging as logging
+import hhnk_research_tools.waterschadeschatter.resources as wss_resources
 from hhnk_research_tools.variables import DEFAULT_NODATA_GENERAL
 from hhnk_research_tools.waterschadeschatter.wss_curves_figures import (
     BergingsCurveFiguur,
@@ -43,6 +45,8 @@ logger = logging.get_logger(__name__)
 
 # Globals
 NAME = "AreaDamageCurves Aggregation"
+
+LANDUSE_CONVERSION_TABLE = hrt.get_pkg_resource_path(wss_resources, "landuse_conversion_table.csv")
 
 # Defaults
 DEFAULT_RAIN = 200  # mm
@@ -74,7 +78,7 @@ class AreaDamageCurvesAggregation:
     aggregate_vector_path: Union[str, Path, None] = None
     aggregate_vector_id_field: Optional[str] = None
     aggregate_vector_layer_name: Optional[str] = None
-    landuse_conversion_path: Union[str, Path, None] = None
+    landuse_conversion_path: Union[str, Path, None] = LANDUSE_CONVERSION_TABLE
 
     def __post_init__(
         self,
@@ -83,15 +87,12 @@ class AreaDamageCurvesAggregation:
 
         if self.dir.output.result_lu_areas.exists():
             self.lu_area_data = pd.read_csv(self.dir.output.result_lu_areas.path, index_col=0)
-            # self.lu_area_data.fid = self.lu_area_data.fid.astype(str)
 
         if self.dir.output.result_lu_damage.exists():
             self.lu_dmg_data = pd.read_csv(self.dir.output.result_lu_damage.path, index_col=0)
-            # self.lu_dmg_data.fid = self.lu_dmg_data.fid.astype(str)
 
         if self.dir.output.result_bu_damage.exists():
             self.bu_dmg_data = pd.read_csv(self.dir.output.result_bu_damage.path, index_col=0)
-            # self.bu_dmg_data.fid = self.bu_dmg_data.fid.astype(str)
 
         self.drainage_areas = self.dir.input.area.load()
         self.drainage_areas[ID_FIELD] = self.drainage_areas[ID_FIELD].astype(str)
@@ -104,8 +105,7 @@ class AreaDamageCurvesAggregation:
             )
             self.field = self.aggregate_vector_id_field
 
-        if self.landuse_conversion_path:
-            self.lu_conversion_table = pd.read_csv(self.landuse_conversion_path)
+        self.lu_conversion_table = pd.read_csv(self.landuse_conversion_path)
 
         self.predicate = DEFAULT_PREDICATE
         # self.time = WSSTimelog(subject=NAME, output_dir=self.dir.post_processing.path)
@@ -255,7 +255,6 @@ class AreaDamageCurvesAggregation:
 
         vol_curves = self.vol_curve[areas_within[ID_FIELD]]
         return pd.Series(vol_curves.sum(axis=1), name=feature[self.field])
- 
 
     def agg_landuse(self, feature, areas_within) -> dict[str, pd.Series]:
         """Sum land use areas data within the given areas."""
@@ -285,12 +284,14 @@ class AreaDamageCurvesAggregation:
     def agg_damage_level_per_ha(self, feature, areas_within) -> dict[str, pd.DataFrame]:
         """Damage level per ha within the given areas."""
         return self.damage_level_per_ha[areas_within[ID_FIELD]]
-        
+
     def agg_damage_level_per_m3(self, feature, areas_within) -> dict[str, pd.DataFrame]:
         """Damage level per m3 within the given areas."""
         return self.damage_per_m3[areas_within[ID_FIELD]]
 
-    def aggregate_rain_curve(self, feature, areas_within, method="lowest_area", mm_rain=DEFAULT_RAIN) -> dict[str, pd.Series]:
+    def aggregate_rain_curve(
+        self, feature, areas_within, method="lowest_area", mm_rain=DEFAULT_RAIN
+    ) -> dict[str, pd.Series]:
         """Different for distribution of rain in the drainage area"""
         if method == "lowest_area":
             output = self.agg_rain_lowest_area(feature, areas_within, mm_rain)
@@ -572,11 +573,11 @@ class AreaDamageCurvesAggregation:
 
     def create_fdla_gpkg(self) -> None:
         """Create GeoPackage file containing drainage area geometries with damage curve data."""
-        
+
         local_file = self.dir.post_processing.peilgebieden.path
         if local_file.exists():
             return
-        
+
         fig_dir = self.dir.post_processing.figures
         ids = self.drainage_areas[ID_FIELD]
 
@@ -722,8 +723,7 @@ class AreaDamageCurvesAggregation:
         return combined
 
     def run(self, aggregation: bool = True, mm_rain: int = DEFAULT_RAIN, create_html: bool = True) -> None:
-        # general data
-        if not self.dir.post_processing.damage_interpolated_curve.exists():    
+        if not self.dir.post_processing.damage_interpolated_curve.exists():
             self.damage_interpolated_curve.to_csv(self.dir.post_processing.damage_interpolated_curve.path)
         if not self.dir.post_processing.volume_interpolated_curve.exists():
             self.vol_interpolated_curve.to_csv(self.dir.post_processing.volume_interpolated_curve.path)
@@ -743,50 +743,57 @@ class AreaDamageCurvesAggregation:
             for _, feature, areas_within in self:
                 name = feature[self.field]
                 agg_dir = self.dir.post_processing.create_aggregate_dir(name)
-                aggregate = self.agg_run(feature, areas_within, mm_rain)
-                agg_damage = self.agg_damage(feature, areas_within)
-                agg_volume = self.agg_volume(feature, areas_within)
-                agg_landuse = self.agg_landuse(feature, areas_within)
-                agg_landuse_dmg = self.agg_landuse_dmg(feature, areas_within)
-                agg_building_dmg = self.agg_buildings_dmg(feature, areas_within)
-                agg_damage_ha = self.agg_damage_level_per_ha(feature, areas_within)
-                agg_damage_m3 = self.agg_damage_level_per_m3(feature, areas_within)
 
-                # aggregations
-                aggregate.index.name = "Volume [m3]"
-                aggregate = aggregate.add_suffix(" [eur]")
-                aggregate.to_csv(agg_dir.aggregate.path)
+                if not agg_dir.selection.path.exists():
+                    self.selection[name].to_file(agg_dir.selection.path)
 
-                # selected geometries
-                self.selection[name].to_file(agg_dir.selection.path)
+                if not agg_dir.aggregate.path.exists():
+                    aggregate = self.agg_run(feature, areas_within, mm_rain)
+                    aggregate.index.name = "Volume [m3]"
+                    aggregate = aggregate.add_suffix(" [eur]")
+                    aggregate.to_csv(agg_dir.aggregate.path)
 
-                agg_damage.index.name = "Peilstijging [m]"
-                agg_damage.name = agg_damage.name + " [euro]"
-                agg_damage.to_csv(agg_dir.agg_damage.path)
+                if not agg_dir.agg_damage.path.exists():
+                    agg_damage = self.agg_damage(feature, areas_within)
+                    agg_damage.index.name = "Peilstijging [m]"
+                    agg_damage.name = agg_damage.name + " [euro]"
+                    agg_damage.to_csv(agg_dir.agg_damage.path)
 
-                agg_volume.index.name = "Peilstijging [m]"
-                agg_volume.name = agg_volume.name + " [m3]"
-                agg_volume.to_csv(agg_dir.agg_volume.path)
+                if not agg_dir.agg_volume.path.exists():
+                    agg_volume = self.agg_volume(feature, areas_within)
+                    agg_volume.index.name = "Peilstijging [m]"
+                    agg_volume.name = agg_volume.name + " [m3]"
+                    agg_volume.to_csv(agg_dir.agg_volume.path)
 
-                agg_landuse.index.name = "Peilstijging [m]"
-                agg_landuse = agg_landuse.add_suffix(" [m2]")
-                agg_landuse.to_csv(agg_dir.agg_landuse.path)
+                if not agg_dir.agg_landuse.path.exists():
+                    agg_landuse = self.agg_landuse(feature, areas_within)
+                    agg_landuse.index.name = "Peilstijging [m]"
+                    agg_landuse = agg_landuse.add_suffix(" [m2]")
+                    agg_landuse.to_csv(agg_dir.agg_landuse.path)
 
-                agg_landuse_dmg.index.name = "Peilstijging [m]"
-                agg_landuse_dmg = agg_landuse_dmg.add_suffix(" [euro]")
-                agg_landuse_dmg.to_csv(agg_dir.agg_landuse_dmg.path)
+                if not agg_dir.agg_landuse_dmg.path.exists():
+                    agg_landuse_dmg = self.agg_landuse_dmg(feature, areas_within)
+                    agg_landuse_dmg.index.name = "Peilstijging [m]"
+                    agg_landuse_dmg = agg_landuse_dmg.add_suffix(" [euro]")
+                    agg_landuse_dmg.to_csv(agg_dir.agg_landuse_dmg.path)
 
-                agg_building_dmg.index.name = "Peilstijging [m]"
-                agg_building_dmg = agg_building_dmg.add_suffix(" [euro]")
-                agg_building_dmg.to_csv(agg_dir.agg_building_dmg.path)
+                if not agg_dir.agg_building_dmg.path.exists():
+                    agg_building_dmg = self.agg_buildings_dmg(feature, areas_within)
+                    agg_building_dmg.index.name = "Peilstijging [m]"
+                    agg_building_dmg = agg_building_dmg.add_suffix(" [euro]")
+                    agg_building_dmg.to_csv(agg_dir.agg_building_dmg.path)
 
-                agg_damage_ha.index.name = "Peilstijging [m]"
-                agg_damage_ha = agg_damage_ha.add_suffix(" [euro/ha]")
-                agg_damage_ha.to_csv(agg_dir.agg_damage_ha.path)
+                if not agg_dir.agg_damage_ha.path.exists():
+                    agg_damage_ha = self.agg_damage_level_per_ha(feature, areas_within)
+                    agg_damage_ha.index.name = "Peilstijging [m]"
+                    agg_damage_ha = agg_damage_ha.add_suffix(" [euro/ha]")
+                    agg_damage_ha.to_csv(agg_dir.agg_damage_ha.path)
 
-                agg_damage_m3.index.name = "Peilstijging [m]"
-                agg_damage_m3 = agg_damage_m3.add_suffix(" [euro/m3]")
-                agg_damage_m3.to_csv(agg_dir.agg_damage_m3.path)
+                if not agg_dir.agg_damage_m3.path.exists():
+                    agg_damage_m3 = self.agg_damage_level_per_m3(feature, areas_within)
+                    agg_damage_m3.index.name = "Peilstijging [m]"
+                    agg_damage_m3 = agg_damage_m3.add_suffix(" [euro/m3]")
+                    agg_damage_m3.to_csv(agg_dir.agg_damage_m3.path)
 
                 agg_dir.create_figure_dir(name)
                 self.create_aggregated_figures(agg_dir, name, feature)
